requirements.txt
Django
djangorestframework
django-environ
django-filter
django-autoslug
django-countries
Pillow
django-phonenumber-field
phonenumbers
psycopg2-binary
flake8
black
isort
djoser
djangorestframework-simplejwt
PyJWT
pytest-django
pytest-factoryboy
Faker
pytest-cov
redis
celery
flower
django-celery-email
argh
PyYAML
watchdog